Ability Locking Gym
Changes the `Initialized` variable on the GymCube to not be private, so that subclasses can use it for init logic as well.
Currently only tests server-only abilities.
The visualisation for the count will need to be updated if the ability also runs on the client, since in that case the ability would just overwrite the local count variable, so it would likely not be a very useful visualisation.